chore(cleanup): move log into internal package
the name of the package is quite bad and it's completly internal to the repo
package, so hide it for now. until the time comes where I can improve it
properly
